apps/agents/pyproject.toml
[project]
name = "axon-agents"
version = "0.0.1"
description = "Axon agent service (LangGraph + FastAPI + multi-LLM)"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.6",
    "uvicorn[standard]>=0.34.0",
    "pydantic>=2.9",
    "httpx>=0.27",
    "tenacity>=9.0",
    "python-dotenv>=1.0",
    "langgraph>=0.2.60",
    "langchain>=0.3.14",
    "langchain-core>=0.3.29",
    "langchain-groq>=0.2.3",
    "langchain-google-genai>=2.0.9",
    "langchain-openai>=0.3.0",
    "langchain-anthropic>=0.3.3",
    "langchain-ollama>=0.2.2",
    "langfuse>=2.57",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.8.4",
    "mypy>=1.14.0",
]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]
